Simple_Agent/requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
